migration/migration.python/requirements.txt
mysql-connector-python==8.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
tqdm==4.66.1